requirements.txt
transformers
torch
sentencepiece
requests
beautifulsoup4
pdfminer.six
python-docx
numpy
streamlit